requirements.txt
django>=5.2
pandas
playwright
python-dotenv
mysqlclient
